Took too much out of tools/mkconfig.c recently.
This commit is cherry-picked from Nuttx, commit 599c796.
This resolves a compilation waring as shown below:
os/include/tinyara/config.h:481:5: warning: "CONFIG_NSOCKET_DESCRIPTORS" is not defined [-Wundef]
^~~~~~~~~~~~~~~~~~~~~~~~~~

When CONFIG_NET is disabled, CONFIG_NSOCKET_DESCRIPTORS is not defined.
But every source code uses it without checking definition because
it is a mandatory CONFIG.
This commit guarantees the definition of CONFIG_NSOCKET_DESCROPTORS.