s5j/gpio: do not place s5jt200_gpio_bank[] on a special section
There is no point in putting the array s5jt200_gpio_bank[] into a
special section for it. Though I do not know any historic background, it
looks very redundant, dirty and inconsistent. Not only that, this makes
us to define some variables in linker script in all target boards that
are based on S5J platform. This is very annoying and has no benefit to
do so.

Change-Id: Idd41b1944491fbf0c3d158a2cfeec7fbd8c094ca
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>