sidk_s5jt200: add comments on initialization functions
For now, initialization stages are very confusing. For example,
platform-/board-/application-specific things are mixed up in a chaos.

As a preparation step for cleaning up, this commit adds missing comments
that describe its purpose of each initialization function. It will be
very helpful to those who are not aware of it. Most of comments were
copied from NuttX.

Change-Id: I9c5f615a306c83160089039d551cd9fbb8f1b0ec
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>