s5j/sflash: allocate sflashdev statically
This commit moves dynamic allocation of sflashdev to static allocation.
There is no reason to have it dynamic given that it is never freed once
initialized. In addition, code lines that manage it dynamically occupies
unnecessary binary footprints.

Here is comparison of binary footprints:

                         text       data        bss         dec
    Before (dynamic)     1344          0         19        1363
    After  (static)      1280         20         15        1315

With this commit, overall footprint has reduced 48bytes. Advantage of
static allocation is clear.

Change-Id: I5d12e455c7ad85ee389d17b1b246695062d30bb9
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>