mtd/progmem: fix incorrect target address calculation
progmem_read/write() is incorrectly calculating the target address,
expecting the offset argument is given in a block number. This is
completely wrong and as a result invalid flash region is accessed.
Byte-oriented read/write interfaces of mtd device accept the target
address in a byte offset, not a block number.

Change-Id: I6b3e731c48b707ecc8274b73723387fb92adf2aa
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>