fs/mtd: enumerate one partition entry at a time
part_procfs_read() tries to enumerate all partitions at a time when it
is requested to read its procfs node. This is problematic when the
caller did not provide enough output buffer. Though the buffer is not
overflowed, users can see only small part as much as it can put in the
buffer. If we enumerate only a single partition entry at a time, this
situation can be prevented, reading the node several times repeatedly.

This commit does not cause any functional differences except the change
of how many entries are enumerated at a time.

Change-Id: I8f152dbddc27caf192510a888422c0bba6a0e76c
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>