fs/smartfs: revert changes by samsung
mksmartfs() accepts a boolean value that specifies how to behave when
the volume is already formatted, to prevent it from being damaged.
However, the way of checking whether the volume is formatted or not is
completely wrong. I never understand what the hell they thought when
they had done this.

Change-Id: Ic778c4ee66c0e66f3b50d97f02113f48117e82bf
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>