s5j/irq: fix incorrect NR_IRQS definition
NR_IRQS is currently defined as 512. This is unnecessarily huge given
that we have 16 SGIs, 16 PPIs, and only 96 SPIs for external interrupt
sources. As a result, the IRQ vector table in kernel, g_irqvector[],
becomes unnecessarily huge, consuming memory spaces.

Here is a comparison table of libkernel.a for each configuration:

    NR_IRQS    .data    .bss     .text     Total
    512        0       3912      25664    29576
    128        0       2376      25664    28040

As you can see, nearly one and half kilobytes of .bss reduces when
NR_IRQS=128.

Change-Id: Ib90aa90677261ad6b2a680a28a8929085905209b
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>