irq/dispatch: add argument ptr to irq_dispatch()
This commit adds a parameter to the interrupt handler for passing
private data to it, so that users can pass additional information to
their callback. This makes drivers that register irq handlers much
simpler than before. It seems to alter quite a lot, but this commit
itself does not inroduce any functional differences at all.

All credits should go to Mark Schulte who wrote the original patch for
NuttX.

Change-Id: Id22891b0f85646e8a2e86a7d4c8463bd8e67cd37
Signed-off-by: Mark Schulte <mark@mjs.pw>
[Shin: back-ported from the latest NuttX version]
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>