drivers/rtc: sync with the lastest NuttX
RTC subsystem (drivers/rtc.c and related header files) in TinyARA seems
to be outdated, so it is hard to write the lower-half driver for it.
This commit updates the RTC subsystem of TinyARA with the lastest NuttX,
which looks well-structured.

This commit mainly updates the source code that is related with RTC and
does not introduce any functional differences.

Change-Id: If622b1c82c645103e43da79d3c48c56e7523d1e7
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>