s5j/timerisr: use RTC as an OS time tick
Using the MCT as an OS systick timer by reserving a local timer seems
overkill. For such purpose, S5J provides the RTC tick counter and using
using it seems more appropriate. This commit configures it as an OS time
tick instead of MCT local timer.

This is a preparation step for redesigning the MCT driver. It does not
introduce any functional differences except changing the tick source.

Change-Id: Ic0f9a20c7d10fee7451c2c0a5857af6c19b07fbf
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>