net: remove NET_NOINTS
NET_NOINTS is a Nuttx uIP legacy knob which determines which
synchronization method is used for network layer. If it is disabled,
interrupts are masked when synchronization is needed. Otherwise,
semaphore is used.

However, in the current implementation, network related function could
not be called in the interrupt context. This means that NET_NOINTS
should be '=y'.

This patch removes NET_NOINTS, and enables it always, so that only
semaphore is used for network layer syncrhonization.

Change-Id: I1dbab2216bd3c1c62b20fc70c653e9129742a163
Signed-off-by: EunBong Song <eunb.song@samsung.com>