arm: s5j: remove up_gpioinitialize()
up_gpioinitialize() was invented while we were working on the BSP for
BCM4390x. Its purpose was for initializing GPIO ports and registering
device nodes for them during the boot sequence, but it was a bad idea.
Available GPIO ports may varies from device to device. This means that
exporting which GPIO ports should be decided at board initialization,
not OS or chip initialization. Having up_gpioinitialize() at the
chip-specific layer is breaking the OS design.

This is why up_gpioinitialize() had been left abandoned for a long time
and finally become an empty function. It is time to remove it
completely. This commit does not make any functional differences.

Change-Id: I0217b377050613502933f4b6ba7b8024ff9937f1
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>