netutils/wifi: rework nvram interfaces
Every boards that are based on S5J should implement Nv_Read/Write/Erase
that provide access NVRAM region WLAN configuration is stored to. The
interface names looks a little bit weird and their prototypes are
passing unnecessary parameters as well. Not only that, it is breaking
the OS design, specifying the physical address where the NVRAM data is
stored to. As flash partition layout might vary in board to board, upper
layer should not have any assumption on it. So, we need to redesign
those interfaces completely new.

This commit renames them to up_wlan_read/write/erase_config() and
changes function prototypes. Unnecessary or board-specific parameters
are dropped. They can be implemented in various manners suited to each
boards, though mtd_config device is being used to store the NVRAM data
in this commit.

Change-Id: I7d50de06f88a4b4ed58c16cc4f3634d5526633eb
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>