s5j: drop s5j_efuse.c
Though it is stored in the internal e-fuse region of S5J chipset, how we
store and retrieve the MAC address on the actual board depends on the
actual board. In other words, managing MAC address is a board-specific
thing. On some board, MAC address might be stored to an external flash.
Thus, up_wlan_get_mac_addr() should be moved to BSP layer, not S5J.

Change-Id: I764e98c1dfb93ebcc227f6c5079731f34d5abd39
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>