build: do not recompile untouched source files
When the object files' build rule whose name includes sub directory
path, in the build rule for object files, "notdir" command causes the
mismatch between the build rule's name and the generated object file
path. So, the build rule for binary regards the object files as the
files not compiled and rebuilds on every build time even if the
object files are not modified.

This patch removes sub directory path from the source code name. And it
also removes "notdir" command from the build rule for object files.

Change-Id: I53874c480d0422ade1c0ecdbbf55f03a1e19419c
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>