builtin: add tash builtin application
The existing tizenRT was made by tash_cmd_install command each time the
application was registered. When you apply this patch, menuconfig will
be able to register to tash as soon as you select it. However, when the
developer registers an application, the following format should be added
in the Makefile.

$(BUILTIN_REGISTRY)$(DELIM)$(APPNAME)_main.bdat: $(DEPCONFIG) Makefile
	$(call REGISTER,$(APPNAME),$(APPNAME)_main,$(THREADEXEC))

	context: $(BUILTIN_REGISTRY)$(DELIM)$(APPNAME)_main.bdat

Change-Id: I09e77e608e3dfeac5e14c7222105ae3404fd0039
Signed-off-by: Junhwan Park <junhwan.park@samsung.com>