s5j/gpio: remove CONFIG_S5J_GPIO
In S5J, GPIO driver is controlling pinmux configuration for other
peripherals. Without it, the entire system cannot work. Thus, having
GPIO or not actually configurable. It should be enabled always. On the
other hands, the lower-half GPIO driver can be excluded from the
compilation when the upper-half driver is disabled.

Change-Id: I81b3d196b4e9091f86f2d9006ee37c020c8df5b9
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>