s5j/rtc: reset BCDMON and BCDDAY explicitly
BCDMON and BCDDAY are set to zero on power-on-reset, which is completely
invalid. This is a kind of hardware errata. Both should be initialized
to 1. This is problematic when we convert RTC time to timespec, because
it leads to underflow. To workaround this problem, we need to explicitly
initialize them to a valid value at boot.

Change-Id: Iab956102cb280211a8a17f9955f23c0e8132a937
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>