s5j: drop to support S5J_DEBUG_BREAK
S5J_DEBUG_BREAK is a dirty hack that was invented to make the target
fall into the infinite loop at boot. It is convenient for us to be able
to hold the target before attaching the debugger. However, we can get
the same result in other various ways. It looks dirty. Let's drop it.

Change-Id: Ia5b30517c03571debd0432d1b24d43590d8d17ae
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>