s5j/serial: treat the debug UART as UART4
In addition to UART0-3, S5J has a special UART port for debugging
purpose. Though there is no reason to git it a special name, it is
called as UARTDBG. Treating it as UART4 simplifies other codes and looks
more consistent. For this reasons, this commit configures the debug UART
using CONFIG_UART4_XXX instead of using CONFIG_UARTDBG_XXX that is
redundantly defined in s5j/Kconfig.

Change-Id: I73cc71e779dc93070983d4d165d9e724b4febd3f
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>