s5j/serial: wait until tx fifo gets empty when init
Characters printed out from eariler stage bootloader is not printed out
or printed in broken. This is because the UART is being initialized
while the Tx transmitter did not complete to send characters in FIFO. It
can be avoided if we can defer the initialization until Tx FIFO gets
empty.

Change-Id: I22bca475aec53cd222f5425417ab8c815fde8251
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>