examples/sensorbd_demo: play a simpler scale instead of a melody
sensorbd_demo contains an example to show the PWM features S5J provides.
It generates pulse of various frequency to make a buzzer sound or
adjusts the brightness of light emitted by LEDs attached to it by
changing the duty cycle.

This commit changes frequencies and duty cycles for buzzer sound in the
sense that it would be better to play a simpler scale rather than to
play some melody.

Change-Id: Ie566ade9d52101a0feca700dc9a11f02fde5590c
Signed-off-by: Bongryul Lee <bongryul.lee@samsung.com>