examples/tls: fix certificates retrieval from SSS
This patch fixes the way the CA and device certificates are retrieved
from the SSS. Before a common ASN1 tag value was searched with 'strstr',
which was failing to properly split up the certificates. Instead we read
the length value of the ASN1 sequence using the field following the
ASN1 tag, which gives us the offset of the next certificate in the chain.

Change-Id: I39c2cc40cd8be527a039c2922b21b97b9ea4a7f2
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>
Signed-off-by: EunBong Song <eunb.song@samsung.com>