s5j/gpio: specify default drive strength values
So far, the driver strength has been ignored or not set when we
configure GPIO pins. As a preparation step for supporting it, this
commit specifies default value of slew rate for each GPIO pins when it
is configured as an alternative function. This commit will not have any
effects until subsequent commit is merged.

Change-Id: Idcf6b1365ed8d9abae40377565b0c1702860fc19
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>