netutils/dhcpd: fix unexpected reset of the network configuration
Stopping the DHCP server is asynchronous and actual release happens
whithin 60 seconds. This causes issues to applications that stop the
DHCP server then change the network configuration before the DHCP
is actually stopped. As a result, the new network configuration is
reset by the DHCP server stop mechanism. This patch disables the
network configuration reset, but should be later replaced by a
cleaner and smarter mechanism.

Change-Id: Ieb8528a15a2d1b46189c0eff5f7f075c78499d03
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>