net/netdev: rework SIOCGIFCONF support
I do not know the background history on netdev_foreach_sync(), but it is
reinventing the wheel which we should avoid as always as possible.
TizenRT already provides a general interface, netdev_foreach() and thus
it does not make any sense to introduce netdev_foreach_sync(), breaking
the OS design. Adding a new API or code in such a way all the time will
make the OS bigger and bigger.

This commit also fixes a buffer overflow issue that might occur when not
enough buffer were passed to ioctl().

Change-Id: Id52c328919efc8d5ba57548af323e3d69335ff9e
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>