stdio: libc: Enable lowstream only for kernel
The invocation of lldbg from the userspace should be
ignored in case of protected build. Because there is no
system call equivalent to up_putc. So for all lldbg calls
from user space, the output is ignored.
This patch avoid direct kernel API calls from the userspace
and handles bot flat and protected builds.

Signed-off-by: Girish K S <ks.giri@samsung.com>