s5j: fix invalid calculations on sector offsets
The size of sector is stored in '.blocksize' field of data structure for
mtd device instance and we should use it when calculating sector
offsets. On the other hand, '.erasesize' holds the size of erase unit
that underlying flash device supports and might be quite different with
sector size. This commit fixes code lines that are confusing them.
Fortunately, those buggy lines have not yet caused any harm, because
both numbers are the same, but this will break the system soon when
start to use different units on them. So let's fix them all.

Change-Id: I2a115fb7dd648257941577f85b9461be81abebde
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>