artik053/progmem: change MTD block size to 256
The size of sector supported by mtd_progmem is 4KB that is a little bit
huge. For example, SSSRW partition stores a number of small certificate
files of hundred bytes, causing significant external fragmentation and
wasting disk spaces. To avoid this problem, we need to keep the sector
size small. This commit changes the internal sector size of progmem to
256 which seems small enough.

Change-Id: I3599f2cdee05633f12fe1a77901adeaeba3c06fc
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>