MTD M2PX: Fix read problem when write/erase is pending
If we READ while a write/erase is pending, the command is ignored
and the write/erase continues. If we dont catch this situation we will
return garbage to the user because the flash will not execute the command.
So READ MUST wait for write completion, and before that, the bus must be
locked since it's a precondition to calling waitwritecomplete().

This patch comes from Nuttx commit id: 2851959d

Change-Id: I2934f7d5d96048a6a97a272d84267ae8551d1198
Signed-off-by: EunBong Song <eunb.song@samsung.com>