libc: add arch-dependent option for ARCH_ROMGETC visibility control
Enabling ARCH_ROMGETC assumes that for the target architecture
there exists implementation of up_romgetc() function. But even when
implementation is missing ARCH_ROMGETC option is still visible
to the user, so the option being selected cause build failure.

The solution is to add upper-level option, which can be selected
in specific architecture configs and make option ARCH_ROMGETC visible.
Otherwise ARCH_ROMGETC will be unvisible to the user.

Signed-off-by: Oleg Lyovin <o.lyovin@partner.samsung.com>