s5j/pm: drop broken PM stuff
PM stuff has been left abandoned since it was introduced just for
proof-of-concept. It is time to clean up. This commit gets rid of those
broken code lines and also invalid comments that were copy-and-pasted
from other NuttX BSP.

This commit does not cause any functional differences because codes that
are being touched were not actually called from anywhere.

Change-Id: Id5b23b2149c7e048b08485ccdd148130ca6579b4
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>

Conflicts:
	os/arch/arm/src/s5j/Kconfig
	os/arch/arm/src/s5j/Make.defs