examples/artik_onboarding: prevent multiple calls to the onboard command
When the onboarding service was started, we used to check in which state
it was before allowing a call to the "onboard" command. However there
was a few seconds window when the service would try to connect with
stored credntials during which a call to the "onboard" command would
cause a crash. This patch adds a new "connecting" state during which
the "onboard" command is not allowed to be called.

Change-Id: I6e1bf54e3d2972d267db268f3fa593583637ad9a
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>