Update sys_arch.c
Fix Message Queue implementation (which is based on Circular Q).
Problem with earlier implementation is that in a particular case when front points to  start of queue and rear points to end of queue, now if new data comes it gets stored at rear (now Queue is full) and rear gets updated as (rear+1%max) which is front. Now if fetch API is called it results in false empty scenario as front == rear evaluates to true while the Queue is full.