iperf: prevent client infinite loop
Sometimes iperf client's write_set fds are never cleared and this can cause
infinite loop. This situation can happen when write_set is not cleared but
test's state is not running because clear write_set is only done in running
state. This patch adds clear write_set fds if test state is TEST_END and th
ere is not cleared write_fds and this can prevent infinite loop.

Change-Id: I3329fc1fa2454793ab2dfcd16eab177294fe3a31
Signed-off-by: EunBong Song <eunb.song@samsung.com>