libc: add support for localtime in date command
Date command did not support localtime (LIBC_LOCALTIME).
This patch gives support for 3 more parameters in date command namely:
Day of Week			[Sunday = Sun, Monday = Mon, and so on...]
Day of Year			[0,365]
Daylight savings time(DST) flag	[positive: DST in effect, 0: DST not in effect, negative: information not availabe]
The parameters are updated based on the RTC support for Localtime.

Signed-off-by: Vidisha Thapa <thapa.v@samsung.com>