Fix for tc_libc_sleep fail issue
The tx and rx tasks are used in tc_libc_unistd_pipe are alive and
wakes up the kernel_tc_main task while testing libc sleep.
So tc of libc sleep is failed.
Using 'waitpid' would solve the above issue, but 'waitpid' depends on CONFIG_SCHED_WAITPID.
And if CONFIG_SCHED_WAITPID is disabled we cannot test tc_libc_unistd_pipe.
So It's better to use pthreads instead of tasks.

Signed-off-by: Lokesh B V <lokesh.bv@partner.samsung.com>