Remove failsafe error setting in pthread_setschedparam.c
Setting the error no at the beginning of logic is not a good practice as it causes
confusion during successful execution of the API.

When API executes successfully, it does not revert the wrong error code set during
beginning of the logic. Thus we should always make a practice to set the error no
only when logic enters the error case.

Signed-off-by: pradeep.ns <pradeep.ns@samsung.com>