S5J I2C feature.
Wrap split I2C transaction as a feature
 controlled by Kconfig option (I2C_ARTIK_EXTENSIONS).
If disabled, the I2C operations conform to the standard.
Once enabled, the i2c_read/i2c_write might not start/end
 i2c transaction, leaving i2c bus in undefined state.
This (split) is how TizenRT works now on Artik platform.

If the extensions are enabled, an ioctl is provided
 to control the I2C operation mode. Enabled extensions
 allow transfers not being transactions (not conforming
 to start-transfer-stop schema).