iotivity : add the library for Tizen RT compatibility in iotivity
This patch adds the library for Tizen RT compatibility in iotivity.
The functions, which are used by iotivity and not defined in Tizen RT,
will be defined at build_common/tizenrt/compatibility/tizenrt_compat.c.

That's why "port_tinyara.c" of apps/examples/iotivity_simpleserver
will not be necessary any more. "port_tinyara.c" will not be removed
for a while in order to support backward compatibility with lower version.

Change-Id: I642f0c83e07e3db59702a12a4b660174d1bcd5eb
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>