Initial port of audio framework code from public git repo.
This commit includes the following pull requests/commit:-
1. https://github.com/Samsung/TizenRT/commit/146b9e3f105c4e15b1d90fc02ef63705311d0ef3
2. https://github.com/Samsung/TizenRT/commit/bba9f1f8265e900131961fdfa4dd5bd9a180db60
3. https://github.com/Samsung/TizenRT/pull/453

Rest of the commits/ pull requests will be added later.

Currently, only compilation has been tested and no actual testing of the code has been done.
The above is the commit message of internal commit. Had to apply this again to external, so that application of further commits is smooth