s5j/serial: fix initialization order
GPIO pinmux configuration should be done before starting to supply PCLK
and EXTCLK to UART when opening it. Otherwise, a noise signal may be
introduced into the receiving FIFO, causing an odd RX interrupt. For the
same reason, this also applies when closing it.

This commit fixes the issue that select() and poll() fails on serial
devices.

Change-Id: I05309a1637574936f53a529d018dd28600b14295
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>