s5j/serial: fix RX availability check
UFSTAT[7:0] holds the number of data in Rx FIFO, but zero when Rx FIFO
is full. Thus, we need additional check on UFSTAT[8] bit to ensure that
the FIFO is empty, if the count is zero. Otherwise, up_rxavailable() may
return 'false' even when Rx FIFO is full, though it is less likely to
happen.

Change-Id: Iffd65bf356dd840e864553daf8aa9a7f04a0368e
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>