Initialize g_logm_overflow_offset value as -1 in logm.c
g_logm_overflow_offset is a variable for saving buffer offset where overflow happened.
So when overflow happens only, It has a value of which the range is from 0 to (buffer size-1).
If not, it should be invalid value like -1.
Initially, logm buffer is empty but logm can decide that overflow happened because this value is 0.
So It is needed to be initialized as invalid value.