fix i2c utc for read/write test
I used the gy30 sensor for i2c test.
gy30 senser adress is 0x23 and it use just 1 byte packet for communication.
So I modified the adress and pakcet size.
And also I modified checking logic the result of read/write function because the return type of read/write is size.