Removing assert failure in audio TCs
In audio utc/itc, readi positive tc gave an assert failure. The failure was due to the last pcm_readi call returning a negative value. The negative value was passed to write function which led to assert failure.
Now an if block has been added which checks the return value of pcm_readi before passing it to the write function.