tools/csvparser: add a limitation of parsing parameters from syscall csv
The g_parm variable which it is a buffer to treat system call parameters
has a maximum value, 16 as a parameter field. But in parse_csvline function,
a maximum value is not checking so that it can cause memory corruption.
Because of above reason, it should have a checking code for maximum.
But, parser can't make a decision whether it is a valid or not. Let's exit
with error.