Prevent reinstallation of an already in-use group key
Supplicant security advisories: https://w1.fi/security/2017-1/
WPA packet number reuse with replayed messages and key reinstallation

[PATCH 2/8] Prevent reinstallation of an already in-use group key

Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

SCSC-Bug-Id: S17SWF-539
Change-Id: I003b3ab7c3a593e6a519f34ea078bf7f9eae9c5a
Signed-off-by: Anoop Nair <anoop.nair@samsung.com>