iotivity: enable to detect error of iotivity immediately in makefile
This patch enables to detect error and stop build immediately when
error occurs in iotivity.

Makefile executes build_iotivity.sh and clean_iotivity.sh scripts in
order to build or clean iotivity. The scripts which executes scons
build system for iotivity doesn't return the result of build. So,
makefile cannot detect immediately if error occurs in iotivity.

This patch makes the scripts return the result and makefile can detect
error of iotivity immediately.

Change-Id: I87296416e56a439dc9c6e61561864e40c3c22f8f
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>