iotivity: unify "build_iotivity.sh" and "clean_iotivity.sh" scripts
This patch unifies "build_iotivity.sh" and "clean_iotivity.sh" scripts
into a single script file, "build_iotivity.sh". The codes of both
script files are almost same. So the codes can be shared in both cases.
When the 1st parameter is "clean", the new unified script will clean
the output files which are generated while building iotivity.

The usage of command is as follows.

  - For building iotivity -> build_iotivity.sh
  - For cleaning iotivity -> build_iotivity.sh clean

Change-Id: Ie3735286c5402368390efb6b27b96a48ae297ab7
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>