examples/iotivity_apps: build only when source codes are modified
This patch makes application build only when source codes are modified.
Because the iotivity_apps's source codes depend on "iotivity_config.h",
the codes should be rebuilt if "iotivity_config.h" is modified.

There are 2 "iotivity_config.h" files in iotivity build environment.
  [1] $(IOTIVITY_OUT_DIR)/resource/c_common/iotivity_config.h
  [2] $(IOTIVITY_OUT_DIR)/include/c_common/iotivity_config.h

[1] file is recreated on every build time, and [2] file is changed only
when its own content is changed. This patch changes the file path to
[2] file's.

Change-Id: I2c477bb2682b8e40011b21ad516d0dd7330ee2ec
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>