artik/onboarding: replace hardcoded certificate by the one from the SE
This patch removes the hardcoded self-signed certificate that was used
by the webserver for the onboarding service REST APIs. It is replaced
by the certificate stored in the secure element. During handshake, the
webserver sends bothe the device certificate and intermediate
certificate, which allows clients to verify the certificate chain with
the ARTIK root CA.

Change-Id: I432fd10a459122b65f7d02cd4e9fa9c79c930e45
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>