arch: checks for the exist of codesinger when buildtime.
In Boards such as ARTIK053s or ARTIK055s, in order to create the correct
image must have to the codesigner tool. But Currently, TizenRT doesn't
check exist of codesinger when buildtime. This patch add config for path
of codesigner in Kconfig and check for exist of codesigner at build time.

Change-Id: I4805c8026740eaf12f1c1e778d70518f540e84ba
Signed-off-by: Jihun Ahn <jhun.ahn@samsung.com>
Signed-off-by: Junhwan Park <junhwan.park@samsung.com>