arm/src: fix data written non-consecutively in Flash
Cache victim can be caused by interrupt during cache flush in write
function in sflash. In this case, data may not be written consecutively
to the flash, and when the power is turned off, it is filled with a
portion of data 0xff.
To prevent this, this patch changed the write address to mirror address
and the mpu attribute to device type.

Change-Id: Ie2261cefc2e138404cf3b4d95abd02e75327bef3
Signed-off-by: Heuisam Kwag <heuisam@samsung.com>