fs/driver: add bit flip detection in smart scan
When power off during erase or write operation, data stored in flash can
not be guaranteed. So this patch adds the following exception handling.
If the headers are erased from the sector of a Flash block containing a
sector that does not contain the crc of the header, add it to the smart
scan to make sure that all data is cleared. If a bit flip occurs in the
entire Flash block, erase the block. If a bit flip occurs only in a
specific sector or if the crc does not match, set the corresponding
sector as release.

Change-Id: Idf80e0ec54b7bb3e73f367a11d6955ef0653a91c
Signed-off-by: Heuisam Kwag <heuisam@samsung.com>