fs/smartfs: add journal verification after write rename transaction
When power off during file rename operation, try fs recover with journal
data in booting time. However, if the crc of the journal data is not
valid, one sector can be set as the first sector of two different files.
If you unlink one of the two files, the chain of another file may be
broken and you may not be able to read it. This patch adds to the
integrity of the written journal data.

Change-Id: I1cc3005ab4b682f05ac39c063d1e0579ff10293a
Signed-off-by: Heuisam Kwag <heuisam@samsung.com>