arch: add board family config
In "Kconfig.board", refer to the file corresponding to each board
name. If the ARTIK05X lineup is more than it is now, this statement
will be longer. So, This commit adds the config of "ARTIK05X_FAMILY",
and this config is selected when the board corresponding to ARTIK05X
is choose.

Change-Id: Ia702d9fc199d93a4fccc81c0cd3cda1a2d4f3966
Signed-off-by: Jihun Ahn <jhun.ahn@samsung.com>