net/lwip: specifiy source addr for DAD global addr
DAD operation have to done even for the global address.
In this case, the link-local address was set to the source address
field as nd6_send_ns() refers only index of link-local address.

This commit is to specify any address(unspecified address) as the
source address for NS when DAD is done.

Change-Id: I3941bdfc29d32136c96e802b0c97b92cdff6f480
Signed-off-by: Changsoon Kim <cs0424.kim@samsung.com>