net/lwip: do not reduce size less than 1280 in case of ipv6 fragment
This patch doesn't let the ipv6 datagram fragment even if MTU is less
than the minimum IPv6 link MTU (1280).

According to "4. Protocol Requirements" in RFC 1981,
"A node may receive a Packet Too Big message reporting a next-hop MTU
that is less than the IPv6 minimum link MTU.  In that case, the node
is not required to reduce the size of subsequent packets sent on the
path to less than the IPv6 minimun link MTU, but rather must include
a Fragment header in those packets"

Change-Id: I40d5cdc27905974cd60934b0a6b74c8f46423fe1
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>