netutils/ntpclient: fix the wrong condition when setting ntp time
This patch fixes the wrong condition when setting ntp time which
is received by ntp server. Originally, this code intends to set
ntp time only when the received timestamp value is valid, But ntp
time is set even if the time is invalid because of wrong condition.

This patch let ntp time be set only when the time is valid.

Change-Id: I07809aacf1e5e52dd5f845ed3334c6abc6302f4c
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>