netutils/dhcpd: fix an issue that dhcpd_stop() takes too much time
Currently, dhcpd_stop() just set a flag to quit loop of dhcpd_run and
wait timeout of select socket. This takes too much time - 10secs as
default timeout for the function to be done. This patch sends trash
packet to the listening socket of server so that the loop can be
terminated immediately.

Change-Id: I651aacd67b200399f1389d107b4c6ce3fd5401eb
Signed-off-by: Siwon Kang <sw0809.kang@samsung.com>