wifi/driver: fix disconnect handling to send event correctly
Driver did not send AP-STA-DISCONNECTED notification to wpa_supplicant
when STA_HAS_LEFT case. This is ironic that slsi_handle_disconnect()
is the point that triggers DISCONNECT event and it is invoked with
reason code - STA_HAS_LEFT. However, the function blocks dispatching
the event because of the reason code. This patch will send event in
any case of STA disconnection.

Change-Id: I9111686e7a42c093c42302eedec0fe4255e396b6
Signed-off-by: Siwon Kang <sw0809.kang@samsung.com>